Separate write_decimal from eosiolib
